Move the Filters into their own file
