+++ src/Mesh/Mesh.jl
@@ -5,5 +5,6 @@ include("Topologies.jl")
 include("Metrics.jl")
 include("Elements.jl")
 include("Grids.jl")
+include("Filters.jl")
 
 end # module
